move local types to shared modules
